+++ src/main/java/insane96mcp/enhancedai/mixin/MeleeAttackGoalMixin.java
@@ -1,6 +1,8 @@
 package insane96mcp.enhancedai.mixin;
 
 import insane96mcp.enhancedai.modules.base.Attacking;
+import insane96mcp.insanelib.base.Feature;
+import net.minecraft.world.InteractionHand;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.PathfinderMob;
 import net.minecraft.world.entity.ai.attributes.AttributeInstance;
@@ -8,7 +10,6 @@ import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.entity.ai.goal.Goal;
 import net.minecraft.world.entity.ai.goal.MeleeAttackGoal;
 import net.minecraft.world.level.pathfinder.Path;
-import net.minecraftforge.common.ForgeMod;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -33,12 +34,20 @@ public abstract class MeleeAttackGoalMixin extends Goal {
 
 	@Shadow private int ticksUntilNextAttack;
 
-	@Inject(at = @At(value = "RETURN"), method = "getAttackReachSqr", cancellable = true)
-	public void onGetAttackReachSqr(LivingEntity livingEntity, CallbackInfoReturnable<Double> callbackInfo) {
-		if (!Attacking.shouldChangeAttackRange())
+	@Shadow protected abstract boolean isTimeToAttack();
+
+	@Shadow protected abstract void resetAttackCooldown();
+
+	@Inject(at = @At(value = "HEAD"), method = "checkAndPerformAttack", cancellable = true)
+	public void getAttackReachSqr(LivingEntity attacked, double distanceSqr, CallbackInfo ci) {
+		if (!Feature.isEnabled(Attacking.class))
 			return;
-		double attackRange = this.mob.getAttributeValue(ForgeMod.ENTITY_REACH.get());
-		callbackInfo.setReturnValue(attackRange * attackRange + livingEntity.getBbWidth());
+		if (Attacking.isWithinMeleeAttackRange(this.mob, attacked) && this.isTimeToAttack() && this.mob.getSensing().hasLineOfSight(attacked)) {
+			this.resetAttackCooldown();
+			this.mob.swing(InteractionHand.MAIN_HAND);
+			this.mob.doHurtTarget(attacked);
+		}
+		ci.cancel();
 	}
 
 	@Inject(at = @At(value = "RETURN"), method = "canContinueToUse", cancellable = true)